Refactor: 카카오 로그인/탈퇴 기능 수정
