+++ src/auth/auth.controller.ts
@@ -18,11 +18,11 @@ export class AuthController {
   ) {}
 
   //* 회원탈퇴
-  @Get('/kakao/withdrawal')
+  @Get('/kakao/deactivate')
   @UseGuards(JwtAuthGuard)
-  async kakaoWithDrawal(@Req() req: CustomRequest) {
+  async kakaoDeactivate(@Req() req: CustomRequest) {
     const userId: number = req.user.userId;
-    return this.authService.kakaoWithDrawal(userId);
+    return await this.authService.kakaoDeactivate(userId);
   }
 
   //! callback url로 지정
@@ -32,37 +32,12 @@ export class AuthController {
     @KakaoUser() kakaoUser: KakaoUserAfterAuth,
     @Res({ passthrough: true }) res
   ): Promise<void> {
-    //* db에 사용자가 있는지 확인
-    let isExistKaKaoUser = await this.authService.findKakaoUser(
-      kakaoUser.snsId
-    );
-    if (!isExistKaKaoUser) {
-      isExistKaKaoUser = await this.authService.createKakaoUser(kakaoUser);
-    }
-
-    //* jwt 발급
-    const jwtPayload = {
-      userId: isExistKaKaoUser.userId,
-    };
-
-    const accessToken = this.jwtService.sign(jwtPayload, {
-      expiresIn: '5h',
-      secret: process.env.JWT_SECRET,
-    });
-
-    console.log(`jwt-accessToken: ${accessToken}`);
-
-    //*! 기존의 쿠키 방식
-    //res.setHeader('Authorization', `Bearer ${accessToken}`);
-    //res.cookie('Authorization', `Bearer ${accessToken}`, {
-    //  httpOnly: false, // JavaScript에서 쿠키에 접근할 수 없도록 설정
-    //  sameSite: 'none',
-    //  secure: true,
-    //  maxAge: 36000000, // 쿠키 만료 시간 설정 (예: 1시간)
-    //});
-
+    //* 1. acessToken 발급
+    const accessToken = await this.authService.kakaoLogin(kakaoUser);
     const redirect_url = `${process.env.CLIENT_URL}/kakaologin?Authorization=${accessToken}`;
     console.log(redirect_url); // 백엔드에서 확인
+
+    //* 2. 프론트로 redirect
     res.redirect(redirect_url);
   }
 }
